Musical voyage including pitch
Music: Getsix - Beautiful Apparation (CC 3.0)
