--- Bullet Time/Assets/ExplosionGun.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: fa251e55eb22c1d41abf5607e350e682
-timeCreated: 1484307211
-licenseType: Pro
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
