--- Bullet Time/Assets/MainLevel.unity.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: afefe222d1804bb448863c6482a76e07
-timeCreated: 1468681098
-licenseType: Pro
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
