+++ Bullet Time/Assets/Scripts/TimeManager.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 0a6025f4c4dd8e14e94e825a5ad07df3
+timeCreated: 1484311561
+licenseType: Pro
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
