+++ Bullet Time/Assets/Sounds/Music/getsix_beautiful-apparation.mp3.meta
@@ -1,0 +1,22 @@
+fileFormatVersion: 2
+guid: c4b8ac6b4c26d7b4b840eee4e5a11156
+timeCreated: 1484519023
+licenseType: Free
+AudioImporter:
+  serializedVersion: 6
+  defaultSettings:
+    loadType: 0
+    sampleRateSetting: 0
+    sampleRateOverride: 44100
+    compressionFormat: 1
+    quality: 1
+    conversionMode: 0
+  platformSettingOverrides: {}
+  forceToMono: 0
+  normalize: 1
+  preloadAudioData: 1
+  loadInBackground: 0
+  3D: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
